# api/docs/swagger.yaml
basePath: /v2
definitions:
...
  errorpkg.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
      moreInfo:
        type: string
      status:
        type: integer
      userMessage:
        type: string
    type: object
...
  router.GroupPatchNameOptions:
    properties:
      name:
        maxLength: 255
        type: string
    required:
    - name
    type: object
  router.GroupRemoveMemberOptions:
    properties:
      userId:
        type: string
    required:
    - userId
    type: object
...
  service.Group:
    properties:
      createTime:
        type: string
      id:
        type: string
      image:
        type: string
      name:
        type: string
      organization:
        $ref: '#/definitions/service.Organization'
      permission:
        type: string
      updateTime:
        type: string
    type: object
  service.GroupCreateOptions:
    properties:
      image:
        type: string
      name:
        maxLength: 255
        type: string
      organizationId:
        type: string
    required:
    - name
    - organizationId
    type: object
  service.GroupList:
    properties:
      data:
        items:
          $ref: '#/definitions/service.Group'
        type: array
      page:
        type: integer
      size:
        type: integer
      totalElements:
        type: integer
      totalPages:
        type: integer
    type: object
...
  service.Organization:
    properties:
      createTime:
        type: string
      id:
        type: string
      image:
        type: string
      name:
        type: string
      permission:
        type: string
      updateTime:
        type: string
    type: object
...
info:
  contact: {}
  title: Voltaserve API
  version: 2.0.0
paths:
...
  /groups:
    get:
      description: List
      operationId: groups_list
      parameters:
      - description: Query
        in: query
        name: query
        type: string
      - description: Organization ID
        in: query
        name: organization_id
        type: string
      - description: Page
        in: query
        name: page
        type: string
      - description: Size
        in: query
        name: size
        type: string
      - description: Sort By
        in: query
        name: sort_by
        type: string
      - description: Sort Order
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.GroupList'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
      summary: List
      tags:
      - Groups
    post:
      consumes:
      - application/json
      description: Create
      operationId: groups_create
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.GroupCreateOptions'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
      summary: Create
      tags:
      - Groups
  /groups/{id}:
    delete:
      consumes:
      - application/json
      description: Delete
      operationId: groups_delete
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
      summary: Delete
      tags:
      - Groups
    get:
      description: Get
      operationId: groups_get
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.Group'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
      summary: Get
      tags:
      - Groups
  /groups/{id}/members:
    delete:
      consumes:
      - application/json
      description: Remove Member
      operationId: groups_remove_member
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.GroupRemoveMemberOptions'
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
      summary: Remove Member
      tags:
      - Groups
    post:
      consumes:
      - application/json
      description: Add Member
      operationId: groups_add_member
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
      summary: Add Member
      tags:
      - Groups
  /groups/{id}/name:
    patch:
      consumes:
      - application/json
      description: Patch Name
      operationId: groups_patch_name
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.GroupPatchNameOptions'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorpkg.ErrorResponse'
      summary: Patch Name
      tags:
      - Groups
...
swagger: "2.0"